fix(openstock): treat envelope latency_ms as f64
Live smoke against the OpenStock runtime (http://192.168.123.104:8040)
showed every /data/fetch response carrying sub-ms `latency_ms` as a
JSON float (e.g. `0.1445915549993515`). The envelope schema declared
it `Option<u64>`, so serde rejected the entire success envelope with
"invalid type: floating point ..., expected u64" — every live fetch
was failing at parse time. The 5 fixture-driven parsers landed in P0.9
never noticed because fixtures used integer latency.

Changed `OpenStockEnvelope::latency_ms` and the flattened
`OpenStockResponse::latency_ms` to `Option<f64>`; updated the one
envelope unit test that asserted on the value.

+++ src/sources/openstock_client.rs
@@ -213,7 +213,7 @@ pub struct OpenStockResponse<T> {
     pub source: String,
     pub artifact_hash: String,
     pub received_at: Option<String>,
-    pub latency_ms: Option<u64>,
+    pub latency_ms: Option<f64>,
 }
 
 impl<T> OpenStockResponse<T> {
